requirements.txt
alembic==1.7.6
anyio==3.5.0
apache-airflow==2.2.3
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.3
apache-airflow-providers-imap==2.2.0
apache-airflow-providers-sqlite==2.1.0
apispec==3.3.2
argcomplete==1.12.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.1
beautifulsoup4==4.10.0
blinker==1.4
boto3==1.21.3
botocore==1.24.3
bs4==0.0.1
cattrs==1.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
configparser==5.2.0
crayons==0.4.0
croniter==1.0.15
cryptography==36.0.1
defusedxml==0.7.1
dill==0.3.4
dnspython==2.2.0
docutils==0.16
email-validator==1.1.3
Flask==1.1.4
Flask-AppBuilder==3.4.4
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.19.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
inflection==0.5.1
iso8601==1.0.2
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.7.1
lockfile==0.12.2
loguru==0.6.0
lxml==4.8.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
numpy==1.22.2
openapi-schema-validator==0.2.3
openapi-spec-validator==0.4.0
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pendulum==2.1.2
prison==0.2.1
psutil==5.9.0
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.11.2
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
rich==11.2.0
s3transfer==0.5.1
selenium==4.1.2
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
spotipy==2.19.0
SQLAlchemy==1.4.31
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
trio==0.20.0
trio-websocket==0.9.2
unicodecsv==0.14.1
Unidecode==1.3.3
urllib3==1.26.8
webdriver-manager==3.5.3
Werkzeug==1.0.1
wsproto==1.0.0
WTForms==2.3.3
youtube-dl==2021.12.17
zipp==3.7.0
